Need most recent AI changes to adapt it to lanes system

+++ src/faction.c
@@ -687,7 +687,7 @@ int faction_isPlayerFriend( int f )
    Faction *faction;
    lua_State *L;
    int errf;
-   double r;
+   int r;
 
    faction = &faction_stack[f];
 
@@ -751,7 +751,7 @@ int faction_isPlayerEnemy( int f )
    Faction *faction;
    lua_State *L;
    int errf;
-   double r;
+   int r;
 
    faction = &faction_stack[f];
